Factoring out common action-duration logic
