+++ pyminion/expansions/seaside.py
@@ -3,11 +3,11 @@ import logging
 from typing import TYPE_CHECKING, Any, List, Optional
 
 from pyminion.core import AbstractDeck, Action, Card, CardType, Treasure
+from pyminion.duration import ActionDuration, BasicNextTurnEffect, RemovePersistentCardsEffect
 from pyminion.effects import (
     AttackEffect,
     EffectAction,
     FuncPlayerGameEffect,
-    PlayerGameEffect,
 )
 from pyminion.expansions.base import copper
 from pyminion.player import Player
@@ -17,88 +17,6 @@ if TYPE_CHECKING:
 
 
 logger = logging.getLogger()
-
-
-class BasicNextTurnEffect(PlayerGameEffect):
-    def __init__(
-        self,
-        name: str,
-        player: Player,
-        card: Card,
-        draw: int = 0,
-        actions: int = 0,
-        money: int = 0,
-        buys: int = 0,
-        discard: int = 0,
-    ):
-        super().__init__(name)
-        self.player = player
-        self.card = card
-        self.draw = draw
-        self.actions = actions
-        self.money = money
-        self.buys = buys
-        self.discard = discard
-
-    def get_action(self) -> EffectAction:
-        if self.draw > 0 and self.discard > 0:
-            return EffectAction.HandAddRemoveCards
-        elif self.draw > 0:
-            return EffectAction.HandAddCards
-        elif self.discard > 0:
-            return EffectAction.HandRemoveCards
-        else:
-            return EffectAction.Other
-
-    def is_triggered(self, player: Player, game: "Game") -> bool:
-        return player is self.player
-
-    def handler(self, player: Player, game: "Game") -> None:
-        if self.draw > 0:
-            player.draw(self.draw)
-
-        player.state.actions += self.actions
-        player.state.money += self.money
-        player.state.buys += self.buys
-
-        if self.discard > 0 and len(player.hand) > 0:
-            if len(player.hand) <= self.discard:
-                discard_cards = player.hand.cards[:]
-            else:
-                discard_cards = player.decider.discard_decision(
-                    prompt=f"Discard {self.discard} card(s) from your hand: ",
-                    card=self.card,
-                    valid_cards=player.hand.cards,
-                    player=player,
-                    game=game,
-                    min_num_discard=self.discard,
-                    max_num_discard=self.discard,
-                )
-                assert len(discard_cards) == self.discard
-
-            for discard_card in discard_cards:
-                player.discard(game, discard_card)
-
-        game.effect_registry.unregister_turn_start_effects(self.get_name(), 1)
-
-
-class RemovePersistentCardsEffect(PlayerGameEffect):
-    def __init__(self, player: Player, cards: List[Card]):
-        super().__init__("Remove Persistent Cards")
-        self.player = player
-        self.cards = cards
-
-    def get_action(self) -> EffectAction:
-        return EffectAction.Other
-
-    def is_triggered(self, player: Player, game: "Game") -> bool:
-        return player is self.player
-
-    def handler(self, player: Player, game: "Game") -> None:
-        for card in self.cards:
-            player.remove_playmat_persistent_card(card)
-
-        game.effect_registry.unregister_turn_start_effects(self.get_name(), 1)
 
 
 class Astrolabe(Treasure):
@@ -143,7 +61,7 @@ class Bazaar(Action):
         )
 
 
-class Caravan(Action):
+class Caravan(ActionDuration):
     """
     +1 Card
     +1 Action
@@ -159,51 +77,8 @@ class Caravan(Action):
             type=(CardType.Action, CardType.Duration),
             draw=1,
             actions=1,
+            next_turn_draw=1,
         )
-
-    def play(self, player: Player, game: "Game", generic_play: bool = True) -> None:
-        self.duration_play(player, game, None, 1, generic_play)
-
-    def multi_play(
-        self,
-        player: Player,
-        game: "Game",
-        multi_play_card: Card,
-        state: Any,
-        generic_play: bool = True,
-    ) -> Any:
-        if state is None:
-            count = 1
-        else:
-            count = int(state) + 1
-
-        self.duration_play(player, game, multi_play_card, count, generic_play)
-
-        return count
-
-    def duration_play(
-        self,
-        player: Player,
-        game: "Game",
-        multi_play_card: Optional[Card],
-        count: int,
-        generic_play: bool = True,
-    ) -> None:
-        if count == 1:
-            persistent_cards: List[Card] = [self]
-            if multi_play_card is not None:
-                persistent_cards.append(multi_play_card)
-
-            for card in persistent_cards:
-                player.add_playmat_persistent_card(card)
-
-            effect = RemovePersistentCardsEffect(player, persistent_cards)
-            game.effect_registry.register_turn_start_effect(effect)
-
-        super().play(player, game, generic_play)
-
-        effect = BasicNextTurnEffect(f"{self.name}: +1 Card", player, self, draw=1)
-        game.effect_registry.register_turn_start_effect(effect)
 
 
 class Cutpurse(Action):
@@ -377,7 +252,9 @@ class NativeVillage(Action):
                 logger.info(f"{player} adds a card to their Native Village mat")
         elif choice == NativeVillage.Choice.GetFromMat:
             mat.move_to(player.hand)
-            logger.info(f"{player} puts {mat_len} card{plural} from their Native Village mat into their hand")
+            logger.info(
+                f"{player} puts {mat_len} card{plural} from their Native Village mat into their hand"
+            )
         else:
             raise ValueError(f"Unknown native village choice '{choice}'")
